+++ javascripts/options.js
@@ -1,5 +1,7 @@
-/*Notations Change*/
 const notations = ["Scientific", "Logarithmic", "Brackets", "Omega", "Cancer", "Zalgo", "Prime", "Blind"];
+const themes = ["Light", "Dark"];
+
+/*Notations Change*/
 function notationChange() {
 	player.options.notationNo = (player.options.notationNo + 1 == notations.length) ? 0 : player.options.notationNo + 1;
 	player.options.notation = notations[player.options.notationNo];
@@ -10,7 +12,20 @@ function targetedNotationChange(notation) {
 	while (player.options.notation != notation) {
 		notationChange();
 	}
-	document.getElementById("notation").innerText = "Notation: " + player.options.notation;
+}
+
+/*Theme Change*/
+function themeChange() {
+	player.options.themeNo = (player.options.themeNo + 1 == themes.length) ? 0 : player.options.themeNo + 1;
+	player.options.theme = themes[player.options.themeNo];
+	document.getElementById("theme").innerText = "Theme: " + player.options.theme;
+	document.getElementById("style").setAttribute("href", "stylesheets/" + player.options.theme.toLowerCase() + ".css");
+}
+
+function targetedThemeChange(theme) {
+	while (player.options.theme != theme) {
+		themeChange();
+	}
 }
 
 /*Save/Load*/