+++ javascripts/prestige/meltdown.js
@@ -44,7 +44,7 @@ function meltdown() {
 
 function simulateMeltdown(tickinterval = 50) {
 	player.meltdown.corium = player.meltdown.corium.add(coriumGainPassive());
-	player.meltdown.amount += meltdownGainPassive();
+	player.meltdown.amount += meltdownGainPassive() * tickInterval / 1000;
 }
 function updateUIMeltdown() {
 	document.getElementById("corium").innerText = notation(player.meltdown.corium);